(#147) clean notifications
- dark mode
- update angular env
- add notification status
- build message in front
- change front for seen notification
- update seen if visible scrollview
- customize profile page
- fix multiple friendRequest
- remove console log
- fix forgotten await on isBlocked

---------

+++ src/nestjs/src/modules/database/friend/service.ts
@@ -75,11 +75,12 @@ export class DBFriendService {
 			meId: me_id,
 			friendId: friend_id,
 		});
-		if (tmp)
-			return await this.friendRepo.delete({
-				meId: me_id,
-				friendId: friend_id,
-			});
-		else throw new NotFoundException("Friend relation not found");
+		if (tmp) return await this.friendRepo.delete({
+			meId: me_id,
+			friendId: friend_id,
+		});
+		else
+			return ;
+
 	}
 }
